Revert "[ENH] add niflow"
