--- .gitignore
@@ -1,10 +1,0 @@
-# python
-__pycache__
-*.egg-info*
-build
-dist
-*.pyc
-
-# vscode related
-.pytest_cache
-.vscode
