+++ environment.yml
@@ -5,7 +5,7 @@ channels:
 - defaults
 dependencies:
 - ca-certificates
-- certifi=2018.10.15=py27_0
+- certifi=2016.2.28=py27_1
 - decorator=4.0.10=py27_0
 - funcsigs=1.0.2=py27_0
 - future=0.15.2=py27_0
@@ -34,7 +34,7 @@ dependencies:
 - pytz=2016.6.1=py27_0
 - rdflib=4.2.1=py27_0
 - requests=2.10.0=py27_0
-- setuptools=41.0.0=py27_0
+- setuptools=23.0.0=py27_1
 - six=1.10.0=py27_0
 - sparqlwrapper=1.7.6=py27_0
 - sqlite=3.13.0=1
