--- niflow-repronim-simple1/doc/README.md
@@ -1,12 +1,0 @@
-# Documentation
-
-This is where you should document your niflow, including any references you would like to cite.
-
-# References
-
-Please cite the following references if you publish work based on the outputs of this niflow:
-
-* Gorgolewski K, Burns CD, Madison C, Clark D, Halchenko YO, Waskom ML, Ghosh SS. (2011).
-  Nipype: a flexible, lightweight and extensible neuroimaging data processing framework in Python.
-  Front. Neuroinform. 5:13.
-  doi:[10.3389/fninf.2011.00013](https://doi.org/10.3389/fninf.2011.00013)
