--- niflow-repronim-simple1/package/setup.cfg
@@ -1,32 +1,0 @@
-[metadata]
-name = niflow-repronim-simple1
-version = attr: niflow.repronim.simple1.__version__
-url = https://github.com/ReproNim/simple_workflow
-author = Yaroslav Halchenko
-author_email = debian@onerussian.com
-description = niflow-repronim-simple1 workflow package
-license = Apache License, 2.0
-classifiers =
-    Development Status :: 2 - Pre-Alpha
-    Environment :: Console
-    Intended Audience :: Developers
-    License :: OSI Approved :: Apache Software License
-    Operating System :: OS Independent
-    Programming Language :: Python :: 2
-    Topic :: Scientific/Engineering
-
-[options]
-packages = find:
-# Place any Python dependencies here
-# install_requires = 
-# Uncomment the following line if your package should distribute any non-Python files
-# include_package_data = True
-
-[options.entry_points]
-console_scripts =
-    niflow-repronim-simple1 = niflow.repronim.simple1.cli:main
-
-## If your workflow requires data to run, uncomment and update the following section:
-# [options.package_data]
-# niflow-repronim-simple1 =
-#     niflow/repronim/simple1/data/*
