+++ startup.sh
@@ -6,4 +6,4 @@ set -eu
 
 export PATH=/opt/repronim/simple_workflow/miniconda/envs/bh_demo/bin:$PATH
 
-$@
+python "$@"
